Regex perf improvement -- like D37948664
Summary: Previously the `label` regex was created using the `regex!` macro; the point was to avoid re-creating the regex each invocation. The problem with this is `Regex` tries to be smart about threading, and the regex is shared across threads. This is problematic because locking on just one regex was a big slow down -- by cloning we avoid re-creation times and also circumvent this locking problem, for a ~33% performance increase on files that were previously especially slow.

Reviewed By: hgoldstein

Differential Revision: D38041816

fbshipit-source-id: 53e26ad43b1b2dec156aebd5a89219edb392da33