Special provision for test/slow/coeffects/fun-param-1.php: Pure unenforced coeffect can't be disambiguated from enforced. 14145/14145
Summary: T126548142: If a user specifies "pure" in their coeffects it'll be stored as a `unenforced_static_coeffect` in the HackCUnit. However the printer prints unenforced and enforced static coeffects identically (`.coeffects_static {}`), so the assembler can't tell that this "pure" is different from `Ctx::Pure`. So this modifies the comparator to allow that ambiguity.

Reviewed By: aorenste

Differential Revision: D38020574

fbshipit-source-id: 845cba545da95eb76829c55d376d2cef36375a91