handle child type struct offset calculation
Summary:
Changed some stuff to allow LdTypeStructureValCns to work with child type structs.

The problem was that given a const string `key` for the field and a potentially non-const `arr`, the instr wouldn't know the `kind` of the type structure thus can't determine if the `key` exists for that particular struct.

**Added FieldsByte to struct:**

The FieldsByte dictates the specific fields that should exist on each bespoke type struct. Each bit is set iff the field(s) at that offset could exist in the struct. The bit that represents each field is defined in the _FIELDS macros as an offset, and these are generally separated according to the different child structs.

Fields in the base TypeStructure always exist and are represented by the offset 0.

Child structs generally contain distinct fields. The only exception is the field generic_types, which can exist for both TSWithClassishTypes and TSWithGenericTypes, so the bit is set for both.

**Then in vasm:**

In irlower-bespoke for `LdTypeStructureValCns`, first access fieldsByte and check that the bit for the field `key` is set. If the bit is set, this means the field could exist (ie. has allocated memory) so proceed to access the actual field using offsets, otherwise return a falsy value.

**Other changes:**
* Moved `kind` to `m_kind` field instead of `m_extra_hi8`. `m_extra_hi8` now stores the `fieldsByte` value for each type of struct
* Changed `getKindOfField` method to `getFieldPair` and now it returns a pair with the DataType of the field and the offset of the field in fieldsByte

Reviewed By: arnabde03

Differential Revision: D37845427

fbshipit-source-id: 7f84e21e5fce371630ffe97546a1784a6e0ecdba