Add beginning of support for extern-worker to HHBBC
Summary:
Add the beginning of support for extern-worker to HHBBC. Right now
HHBBC receives Refs to the UnitEmitters, immediately downloads them,
then processes them as normal. This doesn't accomplish much on its
own, but it lets us restructure HHBBC's input interfaces to handle
Refs directly from HPHPc. Now we can concentrate on changing more of
HHBBC's internals to leverage extern-worker without changing the
interface.

HHBBC's input is now called WholeProgramInput. It stores pairs of
WholeProgramInput::Key and extern_worker Refs to
WholeProgramInput::Value. All of these types are opaque (using pimpl
idiom) to prevent exposing HHBBC internals and allowing for later
changes. This is analagous to php::ProgramPtr (which is now unneeded
and gone).

WholeProgramInput keys and values are constructed from an UnitEmitter
passed to WholeProgramInput::make(). This is meant to be called from
an extern-worker Job. Typically this is just done as part of parsing,
but when HHBBC is run stand-alone, it will load the UnitEmitters from
disk and run the appropriate Jobs. Since WholeProgramInput stores
*Refs* to Value, and not Value itself, we do not have to materialize
the data locally.

Right now Key is empty, and Value just stores the UnitEmitter. This
will change in the future. The idea is that Key will contain metadata
which we want available locally, and most of the unit's data will stay
as a Ref and passed around to various extern-worker Jobs.

There's a bit of include churn caused by removing some unneeded
includes and moving some definitions around.

Reviewed By: edwinsmith

Differential Revision: D38093913

fbshipit-source-id: e38d49d63113353e8e22657e5b3a529211caf6d1