Add GetClsRGProp bytecode and IR
Summary:
This adds a bytecode and IR `GetClsRGProp`, for getting the reified generics property on an object.

The reason why we can't just use `QueryM` to get said property is because `QueryM` uses the current context's class (i.e. the class of `$this`), while what's needed is the class referred to by `self`, `static`, or  `parent`. This is precisely what's done in the `getClsReifiedGenericsProp` function currently, so we can expose and use it.

Reviewed By: oulgen

Differential Revision: D37424312

fbshipit-source-id: 27c8ee55a3b41f8502af9582f899794267044cd5