plumb DeclProvider support for functions, constants, and modules
Summary:
We recently refactored the DeclProvider API so that each kind of
symbol had a dedicated (and typesafe) query method. But only types
are supported since hackc only querys type decls.

This diff adds API plumbing for Functions, Constants, and Modules.
It also fixes a case-sensitivity bug -- type and toplevel function names are not
yet case sensitive so we need to use `eq_ignore_ascii_case()` instead of `==`.

Differential Revision: D38077894

fbshipit-source-id: 9f4bb0f633edb6185b01bf160bd1cf7ebec36b19