Convert memoize enum class labels back to resource on reflection
Summary:
This diff makes is so that an enum class label memoize attribute argument remains a resource after reflection.

As part of dynamically enforced implicit context project, we want to allow enum class labels to be used for memoize attribute argument, i.e.
```
<<__Memoize(#KeyedByIC)>>
```
however, all attributes need to be serializable, since enum class labels do not have general purpose serialization, this does not work. So, we convert the label to a string, which can be serialized, and then later when we reflect on it, we convert it back to an enum class label. This part has been tricky because enum class labels are implemented as resources and what's more is, they use systemlib userland code to deduplicate allocations using memoization. This means that in order to use this deduplication, we need to do a round trip to Hack code.

Reviewed By: paulbiss

Differential Revision: D38001430

fbshipit-source-id: eef5be5f3af577c5358b6581d5ac086135b83f51