Track how often we don't need to re-JIT everything
Summary:
Before we move to the HackC-Translator, we will need a new bcSha1. Currently, we use the hhas printed from HackC. The HackC-Translator skips the step that generates hhas from the Rust HackC Unit.

This diff adds an ODS counter to track how often we are able to re-use a unit and skip re-JITing everything. We can re-use a unit when the bcSha1 matches the bcSha1 of the unit in unit-cache. This applies to both the case where we ReuseUnitsByHash and can lookup a unit by the file hash, and the case where we have to compile a new one, but don't necessarily need to re-JIT everything.

Reviewed By: ricklavoie

Differential Revision: D38154196

fbshipit-source-id: ec07911d558459612c9db7d9ceddf1b8ebf55b1c