Static method dependency fix
Summary:
When writing `parent::foo`, the type-checker first checks if the parent class has a *static* method called `foo`, and if not, if it has a non-static method of that name.

This diff fixes a small bug in the collection of pessimisation dependencies: In the situation above, the pessimisation dependency logic always recorded a dependency from the current method on the static version of `foo`, even if it doesn't exist (because `foo` is non-static). This diff ensures that the pessimisation graph contains a node that correctly reflects the static-ness of `foo` in the parent.

Differential Revision: D37895593

fbshipit-source-id: 5ed58214207fb8985ff6ef11ba9a5c9d41c97d42