produce bespoke type structs from HHBBC
Summary:
### This diff

In HHBBC, store resolved type structures as their bespoke versions when possible and serialize/unserialize the bespoke versions in the repo.

This functionality is hidden behind the same runtime option `EmitBespokeTypeStructures`. I added the option to RepoGlobalData so that repo mode will be able to access. Also added it to a `...hphp_opts` file for testing.

**Main change**

There are 2 places in HHBBC where type structures are being turned into bespokes.

`parse.cpp` - When we load the initial bytecode, some type structures could already be resolved, so we can convert and immediately store them into the unoptimized repo as bespoke.

`hhbbc/type-structure.cpp` - This file resolves all remaining unresolved type structures, and each TS eventually calls `finish()` when resolved. The new method `finishTS()` checks whether the resolved TS could be bespoke and if it is, calls `MakeFromVanilla` to convert it to the bespoke layout. Now calls to `finish()` are split into two cases: for asserting dicts and vecs, and for asserting/converting bespoke type structures. When resolving fields within a type structure that should be dicts or vecs, we should assert those separately using `finish()`. When resolving the actual type structure, we use `finishTS()`.

Added `ArrayKind::BespokeTypeStructure` to the variable serializer so that we can tag bespoke type structures when serializing. They will have a tag of `T` only when used in the `Type::Internal` serialization mode. Bespokes are limited to this mode. In the variable unserializer, we check the tag and create the bespoke array.

**Other changes**

Changed `GetScalarArray` so that bespoke TS will not be escalated into vanilla dicts. Modified ScalarHash functions to handle bespoke arrays: added a bit in the hash to track that array is bespoke, only compare arrays if they are both bespoke or both vanilla.

This allows `GetScalarArray` to also deduplicate bespoke arrays so that we won't serialize duplicate type structs in the repo.

## Next diff

Right now this diff stores the bespoke TS in the repo, but does not use the bespoke version in the JIT, since layout selection relies on logging arrays. Next step is to modify layout selection so that we count bespoke TS when deciding on a layout for sinks.

 ---

**Note:** using type structures in bespoke form adds a lot more noise to logs , specifically `TRACE=bespoke:6` because most iteration accesses requires generic dispatch, which gets logged.  Could be a concern for future debugging?

Reviewed By: paulbiss

Differential Revision: D37423939

fbshipit-source-id: 00a780af98349ea3b2d16b7c55090149415b93ee