Re-org TARGETS for IR w/ autocargo
Summary:
Added autocargo support for the IR crates.  This is necessary to be able to add use of the IR in HackC.

This required a minor re-org of the TARGETS files as autocargo is happier when each library has its own TARGETS file.

In addition autocargo didn't like the crate name "core" so it go renamed to "ir_core".

Also needed to move bc_to_ir/macros.rs into its own directory (w/ its own TARGETS).

Reviewed By: edwinsmith

Differential Revision: D38105883

fbshipit-source-id: d41f12d4fa8077cbc3471526775078fc343273f4