Iterate simplify to remove all unreachable code (take 2)
Summary:
Simplify is the pass that actually makes instructions dealing in Bottoms
Unreachable and prunes them from the unit.  In pruning blocks, it may improve
other types leading to more Bottoms that can be pruned.  This is not just dead
code removal, but a correctness issue.  Much of the irlower codegen is unable
to handle bottom inputs and must be correctly removed prior to irlower.

Simplify also only handled some cases particularly those in which new
instructions were created.  Add a presimplification step that prunes any
instruction using or defing a Bottom.  Furthermore the reverse flow analysis
done to prune blocks that jump to future unreachable blocks was actually broken
so fix that as well.

Reviewed By: ricklavoie

Differential Revision: D38210616

fbshipit-source-id: fb739be8d94626bff23b57c3decfc07baeb7161c