Fix potential non-determinism in RepoGlobalData
Summary:
ConstantFunctions is a vector, so its ordering is
deterministic. However, it is populated from
RuntimeOptions::ConstantFunctions, which is an unordered map. So, the
ordering of ConstantFunctions may not necessarily be
deterministic. Namely, in debug builds, the ordering will be
different. This can cause RE cache misses.

I also noticed a few HHBBC options weren't being serialized. This is
harmless for now because nothing running in an extern-worker job is
reading them, but take this opportunity to fix it.

Reviewed By: oulgen

Differential Revision: D38232965

fbshipit-source-id: 431c36b4d51e7259b5c84fdc2b6b14e72988259a