Add proc macro for rust_to_ocaml attribute
Summary: This attribute macro is intended to be consumed by the rust_to_ocaml codegen tool, so this proc macro doesn't need to do anything other than return the item (with the rust_to_ocaml attribute stripped by rustc). We may want to add validation later so that incorrect use of the attribute emits errors at compile time, but stripping is good enough for now.

Reviewed By: edwinsmith

Differential Revision: D38184812

fbshipit-source-id: 0875dbcf5a5255305a273c74c21cc10f46ee456a