allow jobs not to give up early
Summary:
This starts by changing the signature of MultiWorker.call_stateless to allow the `job` function to return the inputs it hasn't processed.
The way we allow this in the implementation is to have the `merge` function push back those remaining workitems. This is what `merge` in Typing_check_service does under the hood.

Starts by reviewing MultiWorker.mli, then MultiWorker.ml, then Typecheck.ml.

Reviewed By: bobrenjc93

Differential Revision: D38285877

fbshipit-source-id: e49e43e5864d33bec65a3934f3dac8ea41090d51