Derive SHA1 for bytecode hash
Summary:
We need a new bcSha1 for the unit-emitters created via HackC-Translator since we will no longer generate hhas. The HackCUnit from Rust is essentially what we want to hash, since it does not contain the `m_filepath` that the unit-emitter does, which can change but doesn't semantically matter for the unit. Either we do this in C++ as we translate the HackCUnit, or we do this in Rust and pass it to HHVM. I wanted to see what this would look like in Rust, since the types live in rust already, and Rust has a nice use of Traits that easily allows us to separate the visiting of types from the hashing of them, and allows us to statically enforce each type has a `hash()` implementation as the `HackCUnit` type changes.

I initially implemented this using `#[derive(Hash)]` with the SipHasher24, but this only resulted in a 128 bit hash. I think SHA1 is really what we want. The SHA1 hasher requires types implement a trait other than `Hasher`. It requires all types to implement the `Digest` trait, and be able to be converted to a [u8] byte slice `(AsRef<[u8]>)`.

We can't derive this trait, so to get our HackCUnit into a byte array to be hashed, we can just `#[derive(Serialize)]` on our types, and use the `bincode` crate to serialize it into bytes.

Hashing it to a SHA1 gives us a `[u8; 20]`, which we can easily convert to the HPHP::SHA1 type.

The only HackCUnit type that did not derive cleanly was Slice. The Serializer didn't know how to serialize the data field, `const T* data`. Since this type is just C representation of a Rust slice, we can just convert back to a Rust slice and serialize that. I used https://serde.rs/impl-serialize.html for how to serialize sequences.

NOTE: This SHA1 implementation involved serializing, which was something we initially wanted to avoid. Implementing this without serialization could look like defining `AsRef[u8]`  aka convert each type to a byte slice, for each type in the `HackCUnit`. I think this work would mirror the work required to do this in C++, but in Rust, we get the traversal for free. I'm open to suggestions, if there is an easier/better way to do this in C++. I can run some tests for the performance of rust serde to determine whether implementing `AsRef[u8]` is a more viable option.

Reviewed By: aorenste

Differential Revision: D38211492

fbshipit-source-id: 8c2decb66cb0ba3c6c56dd7d1f4e33d419ad1434