Parse HHBBC units in extern-worker jobs
Summary:
Right now we're storing the UnitEmitters in extern-worker during
parsing. HHBBC extracts the UnitEmitters, then parses them into its
own data structures (php::Program).

Modify the pipeline so that the parsing into HHBBC data structures is
done inside the extern-worker job. HHBBC just downloads the data
structures, does a little bit of fixup, then adds them to the
php::Program and proceeds normally. This is the next step of making
HHBBC distributed.

This is mostly just adding serialization support to all of the data
structures in representation.h. This isn't entirely straight-forward
since the data structures can contain pointers to each other. Luckily
(right now) they always point to things in the same unit, so we can
maintain the pointers. This will not be true later on (after trait
flattening), so these will require future modification.

A few things were especially problematic when doing the parsing
distributed. We assign each unit an unique id, and also assign each
function an unique id. The unique unit id is not actually used by
HHBBC, only when the unit is emitted back to the repo. Therefore we
can just defer the generation to there for now. The func id is used
heavily, so just calculate it when downloading the unit and adding it
to the php::Program.

In what is going to be a recurring problem, some of the unit parsing
relies on the runtime being initialized fully. Namely, we need to have
the list of all the native system constants initialized, and might
invoke the runtime to resolve default arguments. None of this works in
extern-worker jobs right now because we deliberately do not do a full
runtime init inside them (because its very slow). This is only a
concern for systemlib units, so an easy solution is to just ensure
those units are always processed locally (not in extern-worker). This
is another solution which will require future revisiting. Add
assertions to ensure we don't try any of this inside an extern-worker
job.

One of the bits of info that the unit parsing also computed was the
list of functions to run the "const pass analysis" on. It turns out
this information is easy to regenerate once we have the entire
php::Program, so just do it there instead of trying to convey it out
of the extern-worker job.

Reviewed By: edwinsmith

Differential Revision: D38267540

fbshipit-source-id: ffb7ce6e7c443b4ec5c288c49984a3dd1d9fb37b