Refactor property code to use MemberLookupContext
Summary:
Rename MethodLookupCallContext to MemberLookupContext, so that we can use it for properties too. Then pass it into the class DeclProp functions.

We leave some TODOs where we need to pass in the full context instead of a class pointer, especially in object-data.h. We will address those callsites in the next diff.

In order to do this safely, I had to change the sentinel value used in the variable unserializer from (Class*)-1 to (Class*)-2, in order to reserve a bit for the Either tag bit. Not sure how to test this beyond what we already have.

Reviewed By: oulgen

Differential Revision: D38024010

fbshipit-source-id: 692199b818b20cb63c62ed22679f80d77079d8b2