Add class and method pointer infrastructure
Summary:
This diff modifies the current hh_single_type_check infrastructure so that users must specify that they are looking for upcasts of some *Function*.

Context: In future diffs, this module will support class pointers and method pointers. By adding this infrastructure early, we will be able to easily modify the CLI so that users can specify whether they are finding Function, Class, or Method pointers. This is especially useful to guard against typos or capitalization errors.

Reviewed By: vassilmladenov

Differential Revision: D38130578

fbshipit-source-id: 15abf7808ba6a0d93c693511a57c2f3065e81c4d