Move function tests into subdirectory
Summary:
This diff moves all of the current tests into their own `function/` directory in preparation for the next diff.

Context: The next diff will add functionality (as well as tests) to support class pointers in this module. Since the class pointer tests require a new CLI command, they will require a separate dune file as well as an HH_FLAGS file. In preparation for this, I moved all of the function files to a separate subdirectory.

Reviewed By: vassilmladenov

Differential Revision: D38062988

fbshipit-source-id: 4da41c33629385dccb7a97510c71854aa284dc51