Fix ARM ABI
Summary:
This diff marks registers d8-d15 as reserved and drops them from the cross-JIT
ABI. D19440142 (https://github.com/facebook/hhvm/commit/bec11da32fd3ae156b457fa426be5028f0427874) added the cross-JIT ABI and included d8-d15 in the registers to
be saved/restored when entering the TC.  However, the push/pop instructions used
to save/restore don't take SIMD registers, so it was incorrect. This diff marks
registers d8-d15 as reserved and drops them from the cross-JIT ABI.  These
registers are a bit tricky in that they're partially callee-saved, thus
requiring them to be saved/restored when entering/exiting the TC if the JIT is
to use them.  This would add overhead to enter/exit the TC and it doesn't sound
worth it since there are 24 other SIMD registers available.  If SIMD register
pressure really becomes an issue, we can revisit this decision to make d8-d15
available for the JIT.

Reviewed By: ricklavoie

Differential Revision: D38363802

fbshipit-source-id: 92593fd4ee0af51c36890e68025c196b78a1681a