Implement ToOcamlRep for rupro decls
Summary:
Make it possible to directly convert a rupro decl to an OCaml value, instead of converting to an oxidized_by_ref value, then converting that to OCaml.

This way, we lose the guarantee of correctness we get from the oxidized_by_ref->OCaml conversion, but we gain a substantial runtime benefit, particularly for reads of values already stored in sharedmem.

Reviewed By: shayne-fletcher

Differential Revision: D38093478

fbshipit-source-id: 979a63771214d413ac9cc34ce3ad736f95375aaa