Call Hierarchy: Get Rid of Declaration Info, Fix File Info
Summary:
Say the definition of function f has in its body a us of function g, which lives in a different file than f.
Previously, when this dependency was discovered, it would be recorded as f explicitly depending on the reference to "g" that lives in its own body, which isn't a declaration. But that reference of g doesn't have a body, so when we asked about g's dependencies, we wouldn't find any.

This diff fixes it so that dependencies are always on the symbol's definition. As a result, the "is_declaration" info that was previously written into the LSP description field is redundant and the associated functions can be eliminated.

Finally, I have fixed one bug in OutgoingCalls here, which also occurs in three other places (IngoingCalls, DepsInBatch, DepsOutBatch) where it will be fixed in a future diff. The bodysymbols function in ServerDepsUtils accepted a parameter "declarations", which was a list of possible symbols it filtered down to the ones properly contained in the body of its other arguments. This name was misleading because it should really be a list of all possible symbols we could want to find in a body, not just declarations.

Instead, we want to filter for symbols of types that could be in the call heirarchy, that is, that could have declarations _somewhere else_. So this diff adds such a function as well.

Reviewed By: panopticoncentral

Differential Revision: D38408950

fbshipit-source-id: 366792655a663ad71e61c720d61c2c026e89df11