Virtualize lambdas and calls
Summary:
It has been requested that Expression Trees allow the option of virtualizing "function types" so that all of the virtual types in an Expression Tree can be captured in a common type hierarchy.

Currently, because function types are simply typed as Hack's `(function(...):...)` type, they cannot fit into a custom type hierarchy.

This diff explores one way of achieving lambda and call virtualization.

1) This diff introduces a `lambdaType` static method on Visitors that is called (virtually) on lambdas in Expression Trees
2) This diff introduces the `__unwrap` method which is called on global function, static methods, and other non instance method invocations of call.

The expectation is that a Visitor implements `lambdaType` as a function with the type `public static function lambdaType<T>(T $_): MyVirtualizedFunctionType<T>` where `MyVirtualizedFunctionType<T>` has a single instance method `__unwrap` that has the signature `public function __unwrap(): T`.

As seen in the test cases, there are a couple of downstream userland effects of this change, type signatures, especially for those in Expression Trees get more complicated. Especially since this is not an opt in - every visitor nearly has to implement a form of function virtualization if we make this change.

Reviewed By: Wilfred

Differential Revision: D32783377

fbshipit-source-id: 1645963093411ac397693cea3d5cbe61154d29f8