Fix is_php_array refinement of supportdyn types
Summary: The refinement code calls get_concrete_supertypes before adding bounds onto the type being refined. This was throwing away `supportdyn` on that type, but we need to keep it to get some `supportdyn` bounds. It's not clear why we need to get the supertypes instead of just adding the type as a bound in the first place and letting the subtyping implementation chase the generics.

Reviewed By: andrewjkennedy

Differential Revision: D38421700

fbshipit-source-id: 5680e93fb44bcd59511662364522a81e0288dfec