Start breaking up php::Program and removing some pointers within HHBBC
Summary:
As we move more HHBBC work to be done in remote jobs, we cannot use
pointers are freely as before. Instead we need more indirect
references to units, funcs, and classes.

We cannot encode the entire Program in one monolithic blob. Instead we
want to encode units, (top-level) funcs, and classes separately. The
unit just includes the misc things which we do not want to encode
separately. So instead of Program just being a vector of units, it is
now a vector of units, funcs, and classes. An unit stores which funcs
and classes it contains by *name* only. Likewise the funcs and classes
store the containing unit by name (as well as references to other
classes and funcs). Using the name is good because its guaranteed to
be unique and is stable across workers.

For the cases where you need an actual pointer, the name needs to be
mapped to the appropriate entity. In most cases we can just use the
Index for this. It already contains mappings of names to entities, so
we can just expose the appropriate lookup functions. The only places
where this doesn't work is in show and debug functions, which might be
called before an Index is constructed. These can just construct the
mappings themselves from a Program.

This diff just breaks up the representation, changes the pointers to
names, and does the lookups where required. The data is still encoded
together and is all available locally. This will start to change in
later diffs.

One other consequence of these changes is that Index now owns the
Program its given. Other places can obtain a const ref to it by
calling an accessor.

While I was at it, I changed a number of functions which take pointers
to refs (since nullptr wasn't an allowed value). I also improved const
correctness in a number of places.

Reviewed By: edwinsmith

Differential Revision: D38442584

fbshipit-source-id: b10360298db63a28df387448ebf110adeeccd663