Pass in record with refactor_mode instead of string to Walker
Summary:
This diff changes the walker so that it takes in `element_info`—a record that includes the class/function name as well as the refactor mode—instead of `element_id`, which is a string containing the class/function name. This is because we will handle methods in the next diff; methods will require additional information beyond just the method name.

Context: The next diff will handle methods. Since method names are not unique (i.e. different classes can have two methods with the same name), we need both the class name and the method name to correctly identify the method that will be refactored.

In the far future, it might also be to consider the refactor mode in the walker, .

Reviewed By: Wilfred

Differential Revision: D38135489

fbshipit-source-id: 0d33bc67f477fb081fcd82e36a4da858892af1ab