Introduce support for class pointers
Summary:
This diff adds the ability to handle class pointers. Now, users can use `hh_single_type_check --enable-sound-dynamic-type --refactor-sound-dynamic simplify Class C` to find the locations of upcasts of class pointers of the Class C.

Class pointers can be upcasted to dynamic, and knowing if a specific class is ever upcasted is useful for a programmer who is writing code with assumptions that a class is not ever used dynamically. Note that since class names and function names are unique, we don't need to tell the walker that we are handing classes instead of functions. Furthermore, the analysis between classes and functions don't need to be different.

Reviewed By: Wilfred

Differential Revision: D38133947

fbshipit-source-id: c83717eb171187551e745fc093344005ea09ad54