Add class dependency information to the WholeProgramInput keys
Summary:
In addition to the name, add class dependency information to the
WholeProgramInput keys. These class dependencies are the classes which
a class extends, uses a trait, or implements. The dependency
information will be needed to perform trait flattening, but we do not
want to load all of the units just to calculate it. It makes more
sense to calculate it as part of parsing, store it in the key (which
is downloaded anyways), and use that to coordinate trait
flattening. Nothing is using the dependency information in the key at
the moment.

Reviewed By: edwinsmith

Differential Revision: D38495082

fbshipit-source-id: 610f15fd556657c0349f3c3b2e570265b0e4c3db