Move loading of HHBBC inputs into Index
Summary:
Move the loading of the HHBBC inputs from whole_program to inside
Index. This is needed because Index will start operating using
extern_worker, so we do not want to immediately load the data. One
consequence of this change is that we no longer have a php::Program
available before the Index is built. This affects some logging (the
parsing output, which has been moved into the parsing step). It also
makes the compressing testing problematic (we do not have a
php::Program to test on). Since this was just for testing and we never
use it anymore, just remove it.

Reviewed By: edwinsmith

Differential Revision: D38495070

fbshipit-source-id: 46e6f26f2d7665647fc794c46774766bab93f297