prepare to migrate Eval.NumReservedSlabs to Eval.NumReservedMBForSlabs
Summary:
Per suggestion from edwinsmith, we want an option that is independent of slab size.

This diff intends to provide a path for migration without much need for synchronization between releases and config updates. The old Eval.NumReservedSlabs still assumes 2M slabs and will be superceded by Eval.NumReservedMBForSlabs. When the new option is set to nonzero, we ignore the old option.

Reviewed By: edwinsmith

Differential Revision: D38505728

fbshipit-source-id: 76b7666dbc042ca05b089557c9cf57e5ab4b0e34