Eliminate supportdynamic in favor of supportdyn
Summary:
This kills the `supportdynamic` type and replaces all references to it with `supportdyn<nonnull | mixed>` (depending on nullability).

The only remaining references to `supportdynamic` are in the the `supportdynamic.hhi` name and `GlobalOptions.tco_experimental_supportdynamic_type_hint` which isn't worth spending time to change the flag string.

This diff exposes a limitation in refinements for `is`, specifically `is_any_array`. I haven't looked into what's happening yet, but I suspect it has to do with the refinement not preserving `supportdyn`-ness when it refines `supportdyn<mixed>`.

Differential Revision: D37950582

fbshipit-source-id: 6456c86369c41aaf9832aebb5cf1a5e34ad02eb5