Enable more like pushing with more generous test for supports dynamic
Summary:
To avoid generating a disjunction during subtyping, the like-pushing rule from D35615041 (https://github.com/facebook/hhvm/commit/93526b65f827e2cfc7c526753134f9ab3b696c51) relies on testing whether a type is *definitely* a dynamic-aware subtype of dynamic. But the implementation of `is_sub_type_for_union_i` and friends is conservative with respect to type variables: it will return `false` for `vec<#1> <:D dynamic` even if `#1` has an upper bound of `supportdyn<mixed>`. This blocks some useful applications of like-pushing, such as that needed to type-check the test below (`push_like_await.good.php`).

In this diff we change the treatment of `#1 <: t` in `simplify_subtype` to simply return `valid` if `t` is literally in the upper bounds of `#1`.

Differential Revision: D38125105

fbshipit-source-id: 2b0f1ff5d0194a7b345c961546cc1439f5564881