store field insertion order in bespoke TS
Summary:
Added field `m_raw_positions` to store the insertion order of fields when creating a bespoke type struct.

m_raw_positions is used to store the order in which fields are added to
the type structure. This order needs to be preserved for iteration and
escalation purposes.

m_raw_positions is divided into 4-bit chunks, and the value in each chunk
represents a field on the type structure. The position value that maps to
each field is defined in the _FIELDS macros. The insertion order is stored
from the lowest to highest bit, so the lowest bit represents the field
that was added first.

This is just a packed integer array for memory efficiency.

```
m_raw_positions  ->
   |  ...  |  0111  |  0001  |  0110  |
                         ^        ^
                         |       lowest bit, field was inserted first
                         |       value is 0110, so inserted field is 'kind'
                         |
                        field inserted second
```

Other changes
* Refactored `GetPosKey` and `GetPosVal` to access fields using insertion order
* `escalateWithCapacity` now escalates fields in order

Reviewed By: paulbiss

Differential Revision: D38105885

fbshipit-source-id: 32376cd07de58140d7dab2016a18812c77254f92